api/requirements.txt
asgiref==3.2.10
autopep8==1.5.3
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.8
django-allauth==0.42.0
django-multiselectfield==0.1.12
django-rest-auth==0.9.5
djangorestframework==3.11.0
idna==2.10
install==1.3.3
oauthlib==3.1.0
pycodestyle==2.6.0
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.9
